+++ CHANGES.md
@@ -2,7 +2,7 @@
 
 ## version 9.2.10
 - Modification pour la version 10 : 
-    - Lors de l'installation, stockage de l'url de base dans l'éventualité de la restauration d'un backup dans un autre arborecence.
+    - Lors de l'installation, stockage de l'url de base dans l'éventualité de la restauration d'un backup dans un autre arborescence.
 
 ## version 9.2.09
 - Corrections :